chore: copy root pages
