+++ vercel-app-playground/src/routes/layout.tsx
@@ -1,35 +1,30 @@
-import { Link } from '@hiogawa/react-server/client';
+import '#/styles/globals.css';
+// import { AddressBar } from '#/ui/address-bar';
+import Byline from '#/ui/byline';
+// import { GlobalNav } from '#/ui/global-nav';
 
-export default function Layout(props: React.PropsWithChildren) {
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
   return (
-    <html>
-      <head>
-        <meta charSet="UTF-8" />
-        <title>React Server Starter</title>
-        <link rel="icon" href="/favicon.ico" />
-      </head>
-      <body>
-        <h3>React Server Starter</h3>
-        <a
-          href="https://github.com/hi-ogawa/vite-plugins/tree/main/packages/react-server"
-          target="_blank"
-        >
-          GitHub
-        </a>
-        <nav>
-          <ul>
-            <li>
-              <Link href="/">Home</Link>
-            </li>
-            <li>
-              <Link href="/use-state">Counter (useState)</Link>
-            </li>
-            <li>
-              <Link href="/server-action">Counter (server action)</Link>
-            </li>
-          </ul>
-        </nav>
-        {props.children}
+    <html lang="en" className="[color-scheme:dark]">
+      <body className="overflow-y-scroll bg-gray-1100 bg-[url('/grid.svg')] pb-36">
+        {/* <GlobalNav /> */}
+
+        <div className="lg:pl-72">
+          <div className="mx-auto max-w-4xl space-y-8 px-2 pt-20 lg:px-8 lg:py-8">
+            <div className="rounded-lg bg-vc-border-gradient p-px shadow-lg shadow-black/20">
+              <div className="rounded-lg bg-black">{/* <AddressBar /> */}</div>
+            </div>
+
+            <div className="rounded-lg bg-vc-border-gradient p-px shadow-lg shadow-black/20">
+              <div className="rounded-lg bg-black p-3.5 lg:p-6">{children}</div>
+            </div>
+            <Byline className="fixed sm:hidden" />
+          </div>
+        </div>
       </body>
     </html>
   );
