--- vercel-app-playground/src/routes/server-action/_action.tsx
@@ -1,11 +1,0 @@
-'use server';
-
-let counter = 0;
-
-export function getCounter() {
-  return counter;
-}
-
-export function changeCounter(formData: FormData) {
-  counter += Number(formData.get('delta'));
-}
