--- vercel-app-playground/src/routes/server-action/_client.tsx
@@ -1,19 +1,0 @@
-'use client';
-
-import { changeCounter } from './_action';
-
-export function Counter({ value }: { value: number }) {
-  return (
-    <form action={changeCounter}>
-      <div>Count: {value}</div>
-      <div>
-        <button name="delta" value={-1}>
-          -1
-        </button>
-        <button name="delta" value={+1}>
-          +1
-        </button>
-      </div>
-    </form>
-  );
-}
