--- vercel-app-playground/src/routes/server-action/page.tsx
@@ -1,10 +1,0 @@
-import { getCounter } from './_action';
-import { Counter } from './_client';
-
-export default async function Page() {
-  return (
-    <div>
-      <Counter value={getCounter()} />
-    </div>
-  );
-}
