--- vercel-app-playground/src/routes/use-state/page.tsx
@@ -1,9 +1,0 @@
-import { Counter } from './_client';
-
-export default function Page() {
-  return (
-    <div>
-      <Counter />
-    </div>
-  );
-}
